removal of some layout pages in favour of handling metadata directly on page.tsx files. getting slug for game pages via serverside data, removing use-client requirement
