--- app/(user)/catchratecalculator/page.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default async function CatchRateCalculator() {
-  return (
-    <main className="flex flex-col justify-center items-center h-screen">
-      <h1>Catch Rate Calculator</h1>
-    </main>
-  );
-}
