--- app/(user)/pokedex/catchratecalculator/loading.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Loading() {
-  return <p>Loading...</p>;
-}
