--- app/(user)/pokedex/catchratecalculator/page.tsx
@@ -1,7 +1,0 @@
-export default async function ShinyRateCalculator() {
-  return (
-    <main className="flex flex-col justify-center items-center h-screen">
-      <h1>Shiny Rate Calculator</h1>
-    </main>
-  );
-}
