--- app/(user)/pokedex/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return <p>Loading...</p>;
-}
