--- app/(user)/pokedex/page.tsx
@@ -1,7 +1,0 @@
-export default async function Pokedex() {
-  return (
-    <main className="flex flex-col justify-center items-center h-screen">
-      <h1>Pokedex</h1>
-    </main>
-  );
-}
