--- app/(user)/pokedex/shinyratecalculator/loading.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Loading() {
-  return <p>Loading...</p>;
-}
