--- app/(user)/pokedex/shinyratecalculator/page.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default async function ShinyRateCalculator() {
-  return (
-    <main className="flex flex-col justify-center items-center h-screen">
-      <h1>Shiny Rate Calculator</h1>
-    </main>
-  );
-}
