--- app/(user)/shinyratecalculator/page.tsx
@@ -1,11 +1,0 @@
-export default async function ShinyRateCalculator() {
-  return (
-    <main className="flex flex-col justify-center items-center h-screen text-white">
-      <h1>Shiny Rate Calculator</h1>
-      <p>
-        Calculate the odds of finding a shiny Pokémon based on game and
-        conditions
-      </p>
-    </main>
-  );
-}
