--- app/(user)/teamplanner/layout.tsx
@@ -1,10 +1,0 @@
-export const metadata = {
-  title: {
-    default: "Team Planner",
-  },
-  description: "Plan your Pokémon teams",
-};
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
